Show only one output type at the same time.
OUTPUT_BINARY takes precedence if specified, and is always output
'raw', without extra information, so it can be piped into other programs.

Otherwise, if specified, OUTPUT_META is shown. If none of these is
specified, the default is to show OUTPUT_ANN (all classes) of the
top PD in the stack.