Pass unitsize to srd_session_send() directly.
Follow an API change in libsigrokdecode where srd_inst_channel_set_all()
no longer gets passed a unitsize, but rather srd_session_send() gets it.

This allows us to avoid having to guess a unitsize based on the number
of channels (which guesses incorrectly in some cases).

This fixes (parts of) bug #352.