Support multiple protocol decoder stacks
Requires libsigrokdecode changes to actually work.
This allows multiple -P options to be specified, each of which is
registered as a full stack with libsigrokdecode.

At this point, there's no way to specify annotations per instance, the
annotations are simply global per decoder (as before) rather than per
instance.

This fixes parts of bug #868.

Signed-off-by: Karl Palsson <karlp@etactica.com>