sigrok-cli: Add Google Trace Event output format (json)
The Chrome browser has a handy event viewer. It accepts JSON files
in the Google Trace Event format and views them in a similar way
to PulseView. The format is a convenient way to share decoded logic
captures. Note that being text based, the filesizes are large, though
Chrome does accept them zipped.

The patch adds a flag which switches the decoder output format:

--protocol-decoder-jsontrace

And an alternative output in show_pd_annotations which formats the
samples appropriately.

An example command line might be:

 sigrok-cli -P i2c:scl=D0:sda=D1 -P i2c:scl=D2:sda=D3 -i test.sr --protocol-decoder-jsontrace > test.json

This would take test.sr, decode two sets of I2C traffic and output as
JSON. Browsing to chrome://tracing from within Chrome, this json file
can then be loaded and viewed.

The patch is currently hard coded to 1MHz samplerate, I'm not
familiar enough with the sigrok code to figure out how to access
the samplerate from show_pd_annotations.