show: add support for -i <fn> --show, provide details on input stream
The combination of an input file spec and --show was not handled before,
and fell through to the case of showing device driver details. Do gather
input stream properties when --show was specified in combination with an
input file, and print those details to stdout.

This implementation does not assume a specific file format, instead the
datafeed callback got modified to watch the packets but not process them
(neither decode nor send data to an output module).

This approach, and the order of option combination checks in main(),
make this implementation work with input files as well as pipes, as well
as with any supported input module or session file, including automatic
format detection.