show: do print floating point results even if value is zero
The availability of a configuration value with floating point data type
was determined by checking its value for being zero. Which prevents the
display of valid data. See the 'offset' in this example:

  $ sigrok-cli -d demo -g Analog --show

Track the current value's availability in a boolean variable. Do print
values even if they are zero.